+++ packages/terra-framework-docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 ## 1.42.0 - (October 20, 2023)
 
 * Added
+  * Added documentation for the new `terra-table` component.
   * Added test to cover enabling zebra striping for the `terra-table` component.
   * Added test for tab focus for scrollable tables for the `terra-table` component.